// packages/sdk/src/wasm/kaspa.d.ts
export enum NetworkType {
  Mainnet = 0,
  Testnet = 1,
  Devnet = 2,
  Simnet = 3,
}
// ...
export enum SighashType {
  All = 0,
  None = 1,
  Single = 2,
  AllAnyOneCanPay = 3,
  NoneAnyOneCanPay = 4,
  SingleAnyOneCanPay = 5,
}
// ...
export interface IUtxoEntry {
    /** @readonly */
    address?: Address;
    /** @readonly */
    outpoint: ITransactionOutpoint;
    /** @readonly */
    amount : bigint;
    /** @readonly */
    scriptPublicKey : IScriptPublicKey;
    /** @readonly */
    blockDaaScore: bigint;
    /** @readonly */
    isCoinbase: boolean;
}




/**
 * Interface defines the structure of a transaction outpoint (used by transaction input).
 * 
 * @category Consensus
 */
export interface ITransactionOutpoint {
    transactionId: HexString;
    index: number;
}
// ...
export interface IScriptPublicKey {
    version : number;
    script: HexString;
}
// ...
export type HexString = string;



/**
 * Color range configuration for Hex View.
 * 
 * @category General
 */ 
export interface IHexViewColor {
    start: number;
    end: number;
    color?: string;
    background?: string;
}

/**
 * Configuration interface for Hex View.
 * 
 * @category General
 */ 
export interface IHexViewConfig {
    offset? : number;
    replacementCharacter? : string;
    width? : number;
    colors? : IHexViewColor[];
}
// ...
export class Address {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
  constructor(address: string);
  static validate(address: string): boolean;
  /**
   * Convert an address to a string.
   */
  toString(): string;
  short(n: number): string;
  readonly version: string;
  readonly prefix: string;
  set setPrefix(value: string);
  readonly payload: string;
}
// ...
export class NetworkId {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
  constructor(value: any);
  toString(): string;
  addressPrefix(): string;
  type: NetworkType;
  get suffix(): number | undefined;
  set suffix(value: number | null | undefined);
  readonly id: string;
}
// ...
export class PublicKey {
  free(): void;
  /**
   * Create a new [`PublicKey`] from a hex-encoded string.
   */
  constructor(key: string);
  toString(): string;
  /**
   * Get the [`Address`] of this PublicKey.
   * Receives a [`NetworkType`] to determine the prefix of the address.
   * JavaScript: `let address = publicKey.toAddress(NetworkType.MAINNET);`.
   */
  toAddress(network: NetworkType | NetworkId | string): Address;
  /**
   * Get `ECDSA` [`Address`] of this PublicKey.
   * Receives a [`NetworkType`] to determine the prefix of the address.
   * JavaScript: `let address = publicKey.toAddress(NetworkType.MAINNET);`.
   */
  toAddressECDSA(network: NetworkType | NetworkId | string): Address;
  toXOnlyPublicKey(): XOnlyPublicKey;
  /**
   * Compute a 4-byte key fingerprint for this public key as a hex string.
   * Default implementation uses `RIPEMD160(SHA256(public_key))`.
   */
  fingerprint(): HexString | undefined;
}
// ...
export class ScriptBuilder {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
  constructor();
  /**
   * Creates a new ScriptBuilder over an existing script.
   * Supplied script can be represented as an `Uint8Array` or a `HexString`.
   */
  static fromScript(script: HexString | Uint8Array): ScriptBuilder;
  /**
   * Pushes the passed opcode to the end of the script. The script will not
   * be modified if pushing the opcode would cause the script to exceed the
   * maximum allowed script engine size.
   */
  addOp(op: number): ScriptBuilder;
  /**
   * Adds the passed opcodes to the end of the script.
   * Supplied opcodes can be represented as an `Uint8Array` or a `HexString`.
   */
  addOps(opcodes: HexString | Uint8Array): ScriptBuilder;
  /**
   * AddData pushes the passed data to the end of the script. It automatically
   * chooses canonical opcodes depending on the length of the data.
   *
   * A zero length buffer will lead to a push of empty data onto the stack (Op0 = OpFalse)
   * and any push of data greater than [`MAX_SCRIPT_ELEMENT_SIZE`](kaspa_txscript::MAX_SCRIPT_ELEMENT_SIZE) will not modify
   * the script since that is not allowed by the script engine.
   *
   * Also, the script will not be modified if pushing the data would cause the script to
   * exceed the maximum allowed script engine size [`MAX_SCRIPTS_SIZE`](kaspa_txscript::MAX_SCRIPTS_SIZE).
   */
  addData(data: HexString | Uint8Array): ScriptBuilder;
  addI64(value: bigint): ScriptBuilder;
  addLockTime(lock_time: bigint): ScriptBuilder;
  addSequence(sequence: bigint): ScriptBuilder;
  static canonicalDataSize(data: HexString | Uint8Array): number;
  /**
   * Get script bytes represented by a hex string.
   */
  toString(): HexString;
  /**
   * Drains (empties) the script builder, returning the
   * script bytes represented by a hex string.
   */
  drain(): HexString;
  /**
   * Creates an equivalent pay-to-script-hash script.
   * Can be used to create an P2SH address.
   * @see {@link addressFromScriptPublicKey}
   */
  createPayToScriptHashScript(): ScriptPublicKey;
  /**
   * Generates a signature script that fits a pay-to-script-hash script.
   */
  encodePayToScriptHashSignatureScript(signature: HexString | Uint8Array): HexString;
  hexView(args?: IHexViewConfig | null): string;
}
/**
 * Represents a Kaspad ScriptPublicKey
 * @category Consensus
 */
export class ScriptPublicKey {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
  constructor(version: number, script: any);
  version: number;
  readonly script: string;
}
// ...
export class XOnlyPublicKey {
  free(): void;
  constructor(key: string);
  toString(): string;
  /**
   * Get the [`Address`] of this XOnlyPublicKey.
   * Receives a [`NetworkType`] to determine the prefix of the address.
   * JavaScript: `let address = xOnlyPublicKey.toAddress(NetworkType.MAINNET);`.
   */
  toAddress(network: NetworkType | NetworkId | string): Address;
  /**
   * Get `ECDSA` [`Address`] of this XOnlyPublicKey.
   * Receives a [`NetworkType`] to determine the prefix of the address.
   * JavaScript: `let address = xOnlyPublicKey.toAddress(NetworkType.MAINNET);`.
   */
  toAddressECDSA(network: NetworkType | NetworkId | string): Address;
  static fromAddress(address: Address): XOnlyPublicKey;
}